+++ tfe/resource_tfe_variable_test.go
@@ -5,8 +5,8 @@ import (
 	"testing"
 
 	tfe "github.com/hashicorp/go-tfe"
-	"github.com/hashicorp/terraform/helper/resource"
-	"github.com/hashicorp/terraform/terraform"
+	"github.com/hashicorp/terraform-plugin-sdk/helper/resource"
+	"github.com/hashicorp/terraform-plugin-sdk/terraform"
 )
 
 func TestAccTFEVariable_basic(t *testing.T) {
@@ -27,6 +27,8 @@ func TestAccTFEVariable_basic(t *testing.T) {
 						"tfe_variable.foobar", "key", "key_test"),
 					resource.TestCheckResourceAttr(
 						"tfe_variable.foobar", "value", "value_test"),
+					resource.TestCheckResourceAttr(
+						"tfe_variable.foobar", "description", "some description"),
 					resource.TestCheckResourceAttr(
 						"tfe_variable.foobar", "category", "env"),
 					resource.TestCheckResourceAttr(
@@ -58,6 +60,8 @@ func TestAccTFEVariable_update(t *testing.T) {
 					resource.TestCheckResourceAttr(
 						"tfe_variable.foobar", "value", "value_test"),
 					resource.TestCheckResourceAttr(
+						"tfe_variable.foobar", "description", "some description"),
+					resource.TestCheckResourceAttr(
 						"tfe_variable.foobar", "category", "env"),
 					resource.TestCheckResourceAttr(
 						"tfe_variable.foobar", "hcl", "false"),
@@ -76,6 +80,8 @@ func TestAccTFEVariable_update(t *testing.T) {
 						"tfe_variable.foobar", "key", "key_updated"),
 					resource.TestCheckResourceAttr(
 						"tfe_variable.foobar", "value", "value_updated"),
+					resource.TestCheckResourceAttr(
+						"tfe_variable.foobar", "description", "another description"),
 					resource.TestCheckResourceAttr(
 						"tfe_variable.foobar", "category", "terraform"),
 					resource.TestCheckResourceAttr(
@@ -122,7 +128,18 @@ func testAccCheckTFEVariableExists(
 			return fmt.Errorf("No instance ID is set")
 		}
 
-		v, err := tfeClient.Variables.Read(ctx, rs.Primary.ID)
+		wsID := rs.Primary.Attributes["workspace_id"]
+		organization, workspace, err := unpackWorkspaceID(wsID)
+		if err != nil {
+			return fmt.Errorf("Unable to unpack workspace ID: %s", wsID)
+		}
+
+		ws, err := tfeClient.Workspaces.Read(ctx, organization, workspace)
+		if err != nil {
+			return fmt.Errorf("Unable to retreive workspace: %s", err)
+		}
+
+		v, err := tfeClient.Variables.Read(ctx, ws.ID, rs.Primary.ID)
 		if err != nil {
 			return err
 		}
@@ -142,6 +159,10 @@ func testAccCheckTFEVariableAttributes(
 
 		if variable.Value != "value_test" {
 			return fmt.Errorf("Bad value: %s", variable.Value)
+		}
+
+		if variable.Description != "some description" {
+			return fmt.Errorf("Bad description: %s", variable.Description)
 		}
 
 		if variable.Category != tfe.CategoryEnv {
@@ -169,6 +190,10 @@ func testAccCheckTFEVariableAttributesUpdate(
 
 		if variable.Value != "" {
 			return fmt.Errorf("Bad value: %s", variable.Value)
+		}
+
+		if variable.Description != "another description" {
+			return fmt.Errorf("Bad description: %s", variable.Description)
 		}
 
 		if variable.Category != tfe.CategoryTerraform {
@@ -199,7 +224,7 @@ func testAccCheckTFEVariableDestroy(s *terraform.State) error {
 			return fmt.Errorf("No instance ID is set")
 		}
 
-		_, err := tfeClient.Variables.Read(ctx, rs.Primary.ID)
+		_, err := tfeClient.Variables.Read(ctx, rs.Primary.Attributes["workspace_id"], rs.Primary.ID)
 		if err == nil {
 			return fmt.Errorf("Variable %s still exists", rs.Primary.ID)
 		}
@@ -222,6 +247,7 @@ resource "tfe_workspace" "foobar" {
 resource "tfe_variable" "foobar" {
   key          = "key_test"
   value        = "value_test"
+  description  = "some description"
   category     = "env"
   workspace_id = "${tfe_workspace.foobar.id}"
 }`
@@ -240,6 +266,7 @@ resource "tfe_workspace" "foobar" {
 resource "tfe_variable" "foobar" {
   key          = "key_updated"
   value        = "value_updated"
+  description  = "another description"
   category     = "terraform"
   hcl          = true
   sensitive    = true